refactor: split utils.ts into assert-utils and event-utils
